tfp-foundation-protocol/requirements.txt
pytest==8.3.5
fastapi==0.115.12
uvicorn==0.34.0
httpx==0.28.1
python-ndn==0.5.1
pytest-asyncio==0.23.8
websockets>=11.0
redis>=5.0.0
prometheus-client>=0.20.0
pyyaml>=6.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0
psycopg2-binary>=2.9.9
